chore: rearrange header
