+++ common/autoware_point_types/include/autoware_point_types/types.hpp
@@ -17,6 +17,8 @@
 
 #include <point_cloud_msg_wrapper/point_cloud_msg_wrapper.hpp>
 
+#include <pcl/point_types.h>
+
 #include <cmath>
 #include <tuple>
 
@@ -39,6 +41,17 @@ struct PointXYZI
     return float_eq<float>(p1.x, p2.x) && float_eq<float>(p1.y, p2.y) &&
            float_eq<float>(p1.z, p2.z) && float_eq<float>(p1.intensity, p2.intensity);
   }
+};
+
+enum ReturnType : uint8_t {
+  INVALID = 0,
+  SINGLE_STRONGEST,
+  SINGLE_LAST,
+  DUAL_STRONGEST_FIRST,
+  DUAL_STRONGEST_LAST,
+  DUAL_WEAK_FIRST,
+  DUAL_WEAK_LAST,
+  DUAL_ONLY,
 };
 
 struct PointXYZIRADRT
@@ -76,5 +89,11 @@ using PointXYZIRADRTGenerator = std::tuple<
   field_return_type_generator, field_time_stamp_generator>;
 
 }  // namespace autoware_point_types
+
+POINT_CLOUD_REGISTER_POINT_STRUCT(
+  autoware_point_types::PointXYZIRADRT,
+  (float, x, x)(float, y, y)(float, z, z)(float, intensity, intensity)(std::uint16_t, ring, ring)(
+    float, azimuth, azimuth)(float, distance, distance)(std::uint8_t, return_type, return_type)(
+    double, time_stamp, time_stamp))
 
 #endif  // AUTOWARE_POINT_TYPES__TYPES_HPP_
